Move signature reading out of config.py into `_annotations`
Deciding which parameters ask for a config had grown to a third of
config.py, in front of the Config API rather than beside it. It is a
self-contained job — one entry point in, and only the marker class out —
so it moves to a module of its own, which is also where the two rules it
answers to can be stated once: resolve an annotation only in the namespace
that wrote it, and never fail or run user code to find out what it means.
They were previously implicit across six docstrings.

`Config` is passed to `declarations_for()` rather than imported there, so
reading signatures has nothing to say about what a config means, and the
cache stays with the lookup it belongs to.

Three cleanups carried along:

- `_lazy_slots()` built two sets that were consumed one line later in the
  same call; `_instantiate_internal` asks per argument instead.
- "Lazy" meant both "receives the config unresolved" and "classified on
  demand". The names now say which: `_config_parameters(target)`, and
  `wants_config.positional(index)` at the call site.
- The state threaded through the annotation walk (`_seen`, `_bound`) was
  two underscore-prefixed parameters that recursion had to remember to
  pass on. It is a `_Resolution` carrying its own `wants_marker()`.
